refactor: implementing District logic in PropertyService and custom exceptions

+++ src/main/java/br/com/group9/desafio_quality/controller/PropertyController.java
@@ -9,6 +9,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -46,8 +47,8 @@ public class PropertyController {
     }
 
     @GetMapping("/property/{id}/totalValue")
-    public ResponseEntity<Double> getTotalValue(@PathVariable Long id) {
-        Double totalValue = propertyService.getTotalValueByPropertyId(id);
+    public ResponseEntity<BigDecimal> getTotalValue(@PathVariable Long id) {
+        BigDecimal totalValue = propertyService.getTotalValueByPropertyId(id);
         return new ResponseEntity<>(totalValue, HttpStatus.OK);
     }
 }
